wrong int types in ResponseData, LimitResponseData and CostResponseDa…

+++ src/DTO/CostResponseData.php
@@ -176,7 +176,7 @@ class CostResponseData extends DataTransferObject
     /**
      * Usage associated with the request
      *
-     * @var int|null
+     * @var float|null
      */
-    public ?int $usage;
+    public ?float $usage;
 }
